Refactor
* Retain native calls at the UI layer
* Move `TryGetExactPath` to its only call site
* Add a controller for `FormFileHistory` form that orchestrates all non-UI logic
* Move tests
